-updated ccache package to v3.6, , release date: 2019-01-14
-added pvs-studio-analyzer option "--disableLicenseExpirationCheck"
-refactored .gitattributes to better handle line endings
-correct CXXFLAGS for leveldb (-std=c++0x -> -std=c++17)
-added -j3 (job count) parameter while running PVS Studio Analyzer in CircleCI to match Medium+ resource class

+++ src/wallet/rpcwallet.cpp
@@ -4386,7 +4386,7 @@ Examples:
             nSaplingNoteCount += wtx.second.mapSaplingNoteData.size();
     }
     UniValue ret(UniValue::VOBJ);
-    ret.pushKV("sapling", nSaplingNoteCount);
+    ret.pushKV("sapling", static_cast<int64_t>(nSaplingNoteCount));
 
     return ret;
 }
